+++ Mozu.Api/Contracts/Content/Facet.cs
@@ -12,19 +12,10 @@ using System;
 
 namespace Mozu.Api.Contracts.Content
 {
-		///
-		///	Properties of the facet used to retrieve documents.
-		///
 		public class Facet
 		{
-			///
-			///The number of facet results.
-			///
-			public long Count { get; set; }
+			public int Count { get; set; }
 
-			///
-			///The name of the facet.
-			///
 			public string Name { get; set; }
 
 		}
