+++ Mozu.Api/Contracts/Core/UserCollection.cs
@@ -13,19 +13,10 @@ using System.Collections.Generic;
 
 namespace Mozu.Api.Contracts.Core
 {
-		///
-		///	Collection of users and user details.
-		///
 		public class UserCollection
 		{
-			///
-			///The number of results listed in the query collection, represented by a signed 64-bit (8-byte) integer. This value is system-supplied and read-only.
-			///
-			public int TotalCount { get; set; }
+			public long TotalCount { get; set; }
 
-			///
-			///An array list of objects in the returned collection.
-			///
 			public List<User> Items { get; set; }
 
 		}
