+++ Mozu.Api/Contracts/ProductAdmin/ProductCollection.cs
@@ -13,31 +13,16 @@ using System.Collections.Generic;
 
 namespace Mozu.Api.Contracts.ProductAdmin
 {
-		///
-		///	The container for a set of products.
-		///
 		public class ProductCollection
 		{
-			///
-			///The number of pages returned based on the startIndex and pageSize values specified. This value is system-supplied and read-only.
-			///
-			public int PageCount { get; set; }
+			public long PageCount { get; set; }
 
-			///
-			///Specifies the number of results to display on each page when creating paged results from a query. The maximum value is 200.
-			///
 			public int PageSize { get; set; }
 
 			public int StartIndex { get; set; }
 
-			///
-			///The number of results listed in the query collection, represented by a signed 64-bit (8-byte) integer. This value is system-supplied and read-only.
-			///
-			public int TotalCount { get; set; }
+			public long TotalCount { get; set; }
 
-			///
-			///An array list of objects in the returned collection.
-			///
 			public List<Product> Items { get; set; }
 
 		}
