+++ Mozu.Api/Contracts/ProductRuntime/ProductSearchResult.cs
@@ -13,36 +13,18 @@ using System.Collections.Generic;
 
 namespace Mozu.Api.Contracts.ProductRuntime
 {
-		///
-		///	The result of a product search.
-		///
 		public class ProductSearchResult
 		{
-			///
-			///The number of pages returned based on the startIndex and pageSize values specified. This value is system-supplied and read-only.
-			///
-			public int PageCount { get; set; }
+			public long PageCount { get; set; }
 
-			///
-			///Specifies the number of results to display on each page when creating paged results from a query. The maximum value is 200.
-			///
 			public int PageSize { get; set; }
 
 			public int StartIndex { get; set; }
 
-			///
-			///The number of results listed in the query collection, represented by a signed 64-bit (8-byte) integer. This value is system-supplied and read-only.
-			///
-			public int TotalCount { get; set; }
+			public long TotalCount { get; set; }
 
-			///
-			///The facets applied to index products in the product search result.
-			///
 			public List<Facet> Facets { get; set; }
 
-			///
-			///An array list of objects in the returned collection.
-			///
 			public List<Product> Items { get; set; }
 
 		}
